Finalização google maps

+++ src/main/java/br/com/trixmaps/model/Location.java
@@ -22,10 +22,10 @@ public class Location extends AbstractEntity{
 
 	
 	@Column(nullable=false)
-	private Integer latitude;
+	private Double latitude;
 	
 	@Column(nullable=false)
-	private Integer longitude;
+	private Double longitude;
 	
 	@Column(length=30, nullable=false)
 	private String name;
@@ -42,19 +42,19 @@ public class Location extends AbstractEntity{
 		super();
 	}
 
-	public Integer getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Integer latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public Integer getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Integer longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
